Start adding unit tests for the ResourceForks.
These tests will help test the core parsing functionality of the
resource forks to ensure that they remain functional, correct and
reliable.

There needs to be some code reorganisation and to bring the resourcefork
structures out into a header file marked "private" so that the tests can
easily see those definitions and functions.